Первая часть ДЗ 5ого раздела 'Модуляция.'
